bug 1153288 - Import drf-json-api code, step 1
The v1 API uses JSON API release candidate 2 format, partially
implemented by drf-json-api.  That code base is moving on to JSON API
v1.0, so we'll need to preserve the old format until v1 goes away.

Step 1 is copying the relevant code without much modification.  Coverage
is no longer 100%, because a lot of the drf-json-api functionaility is
unused or explicitly replaced.