bug 1208913 - Add 'child_pages' query parameter
When a view_feature include the query parameter child_pages=1, it
includes the old behaviour of returning a paginated set of all child
features.  If it is set to 0 or omitted, then only the row-level child
features are returned, and they are not paginated. This more closely
matches what is expected for constructing MDN tables.

This also adds a new item, meta.compat_table.child_pages, to reflect
what setting was used.